unguarded all atributes, added JWT to all routes so other users cannot change information on routes

+++ back-end/app/Http/Controllers/MainTaskController.php
@@ -6,6 +6,7 @@ namespace App\Http\Controllers;
 use App\Models\MainTask;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Http\Request;
+use Tymon\JWTAuth\Facades\JWTAuth;
 use function Laravel\Prompts\error;
 
 class MainTaskController extends Controller
@@ -13,11 +14,12 @@ class MainTaskController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index(string $id)
+    public function index()
     {
+        $userId = JWTAuth::parseToken()->authenticate()->id;
 
-        return MainTask::with(['group', 'users'])->whereHas('users', function ($query) use ($id) {
-            $query->where('users.id', $id);
+        return MainTask::with(['group', 'users'])->whereHas('users', function ($query) use ($userId) {
+            $query->where('users.id', $userId);
         })->get();
     }
 
@@ -26,10 +28,14 @@ class MainTaskController extends Controller
      */
     public function create(request $request)
     {
-        if (!$request->title || !$request->deadline || !$request->ai_file) {
-            return response(['error' => 'you are stupid'], 404);
-        }
         try {
+
+            if (!$request->title || !$request->deadline || !$request->ai_file) {
+                return response(['error' => 'you are stupid'], 404);
+            }
+// todo aanpassen naar admin ipv user_id
+            $userId = JWTAuth::parseToken()->authenticate()->id;
+
             $mainTask = new MainTask([
                 'title' => $request->title,
                 'deadline' => $request->deadline,
@@ -40,7 +46,7 @@ class MainTaskController extends Controller
             ]);
             $mainTask->save();
 
-            $mainTask->users()->attach($request->user_id, [
+            $mainTask->users()->attach($userId, [
                 'level' => $request->level ?? "beginner",
                 'progress' => $request->progress ?? 0,
                 'score' => $request->score ?? null,
@@ -58,7 +64,13 @@ class MainTaskController extends Controller
      */
     public function show(string $id)
     {
-        return MainTask::query()->findOrFail($id);
+        $userId = JWTAuth::parseToken()->authenticate()->id;
+        $mainTask = MainTask::query()->findOrFail($id);
+        if ($mainTask->user_id == $userId) {
+            return $mainTask;
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
+        }
     }
 
     /**
@@ -71,25 +83,32 @@ class MainTaskController extends Controller
             if (!$request->title || !$request->deadline || !$request->ai_file) {
                 return response(['error' => 'you are stupid'], 404);
             }
+
             $mainTask = MainTask::query()->findOrFail($id);
+            $userId = JWTAuth::parseToken()->authenticate()->id;
+            // todo aanpassen naar admin ipv user_id
+            if ($mainTask->user_id === $userId) {
+                $mainTask->update([
+                    'title' => $request->title ?? $mainTask->title,
+                    'deadline' => $request->deadline ?? $mainTask->deadline,
+                    'description' => $request->description ?? $mainTask->description,
+                    'ai_file' => $request->ai_file ?? $mainTask->ai_file,
+                    'group_id' => $request->group_id ?? $mainTask->group_id,
 
-            // is kinda redundant, but we'll leave it there for safety
-            $mainTask->update([
-                'title' => $request->title ?? $mainTask->title,
-                'deadline' => $request->deadline ?? $mainTask->deadline,
-                'description' => $request->description ?? $mainTask->description,
-                'ai_file' => $request->ai_file ?? $mainTask->ai_file,
-                'group_id' => $request->group_id ?? $mainTask->group_id,
+                ]);
+                $mainTask->save();
 
-            ]);
-            $mainTask->save();
+                $mainTask->users()->updateExistingPivot($userId, [
+                    'level' => $request->level ?? $mainTask->level,
+                    'progress' => $request->progress ?? $mainTask->progress,
+                    'score' => $request->score ?? $mainTask->score,
+                ]);
+                return $mainTask;
+            } else {
+                return response()->json(['error' => 'you are not authorized'], 403);
+            }
 
-            $mainTask->users()->updateExistingPivot($request->user_id, [
-                'level' => $request->level ?? $mainTask->level,
-                'progress' => $request->progress ?? $mainTask->progress,
-                'score' => $request->score ?? $mainTask->score,
-            ]);
-            return $mainTask;
+
         } catch (ModelNotFoundException $e) {
             return response(['error' => $e], 500);
         }
@@ -101,8 +120,16 @@ class MainTaskController extends Controller
      */
     public function delete(string $id)
     {
+
+// todo aanpassen naar admin ipv user_id
+        $userId = JWTAuth::parseToken()->authenticate()->id;
         $mainTask = MainTask::query()->findOrFail($id);
-        $mainTask->delete();
-        return $mainTask;
+
+        if ($mainTask->user_id === $userId) {
+            $mainTask->delete();
+            return $mainTask;
+        } else {
+            return response()->json(['error' => 'you are not authorized'], 403);
+        }
     }
 }
